Tasks up to 4.23* - final commit to part 4.

+++ part4/blog-list/src/routes/blogs/index.ts
@@ -1,32 +1,38 @@
 import express, { Request, Response } from 'express'
+import { JsonWebTokenError } from 'jsonwebtoken'
 
 import { IApiResponse } from '../../interfaces/IApiResponse'
-import { IBlogDoc } from '../../interfaces/IBlog'
+import { IBlogClient } from '../../interfaces/IBlog'
 import { Blog } from '../../models/blog'
+import { tokenExtractor, userExtractor } from '../../helpers/auth'
 
 const blogs = express.Router()
 
 blogs.get(
   '/',
-  async (req: Request, res: Response<IApiResponse<IBlogDoc[]>>) => {
-    const blogs = await Blog.find({})
+  async (req: Request, res: Response<IApiResponse<IBlogClient[]>>) => {
+    const blogs = await Blog.find({}).populate('user', {
+      username: 1,
+      name: 1,
+      id: 1,
+    })
 
     return res.json({
-      data: blogs,
+      data: blogs.map((blog) => blog.toJSON()),
     })
   }
 )
 
 blogs.get(
   '/:id',
-  async (req: Request, res: Response<IApiResponse<IBlogDoc>>) => {
+  async (req: Request, res: Response<IApiResponse<IBlogClient>>) => {
     const { id } = req.params
 
     const blog = await Blog.findOne({ _id: id })
 
     if (blog) {
       return res.json({
-        data: blog,
+        data: blog.toJSON(),
       })
     } else {
       return res.status(404).json({
@@ -38,25 +44,65 @@ blogs.get(
 
 blogs.delete(
   '/:id',
+  tokenExtractor,
+  userExtractor,
   async (req: Request, res: Response<IApiResponse<null>>) => {
-    await Blog.findByIdAndRemove(req.params.id)
+    const { token, decodedToken, user } = req
 
-    return res.status(204).json({})
+    if (!token || !decodedToken) {
+      throw new JsonWebTokenError('Invalid or missing token.')
+    }
+
+    if (!user) {
+      return res.status(401).json({ error: { message: 'User not found.' } })
+    }
+
+    const blog = await Blog.findById({ _id: req.params.id })
+
+    if (!blog) {
+      return res.status(400).json({ error: { message: 'Blog not found.' } })
+    }
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(req.params.id)
+
+      return res.status(204).json({})
+    }
+
+    return res.status(401).json({ error: { message: 'Not allowed.' } })
   }
 )
 
-blogs.post('/', async (req: Request, res: Response<IApiResponse<IBlogDoc>>) => {
-  const blog = new Blog(req.body)
-  const savedBlog = await blog.save()
+blogs.post(
+  '/',
+  tokenExtractor,
+  userExtractor,
+  async (req: Request, res: Response<IApiResponse<IBlogClient>>) => {
+    const { token, decodedToken, user } = req
 
-  return res.json({
-    data: savedBlog,
-  })
-})
+    if (!token || !decodedToken) {
+      throw new JsonWebTokenError('Invalid or missing token.')
+    }
+
+    if (!user) {
+      return res.status(401).json({ error: { message: 'User not found.' } })
+    }
+
+    const blog = new Blog({ ...req.body, user: user._id })
+    const savedBlog = await blog.save()
+
+    user.blogs = user.blogs ? user.blogs.concat(savedBlog._id) : [savedBlog._id]
+    await user.save()
+
+    return res.json({
+      data: savedBlog.toJSON(),
+    })
+  }
+)
 
 blogs.patch(
   '/:id',
-  async (req: Request, res: Response<IApiResponse<IBlogDoc>>) => {
+  async (req: Request, res: Response<IApiResponse<IBlogClient>>) => {
     const { id } = req.params
     await Blog.findByIdAndUpdate(id, req.body)
